+++ squadcastv1/src/lib/config.ts
@@ -61,8 +61,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.2.0",
-  genVersion: "2.729.2",
+  sdkVersion: "0.3.0",
+  genVersion: "2.730.0",
   userAgent:
-    "speakeasy-sdk/typescript 0.2.0 2.729.2 1.0.0 @solarwinds/squadcast-sdk-typescript",
+    "speakeasy-sdk/typescript 0.3.0 2.730.0 1.0.0 @solarwinds/squadcast-sdk-typescript",
 } as const;